Refacto : Move sound service to adapter secondary
