--- frontend/src/app/services/sound/sample.ts
@@ -1,4 +1,0 @@
-export interface Sample {
-  fileName: string;
-  sample?: AudioBuffer;  // Optional since it's marked as undefined in the class
-}
